`beforeEnd:` in http response handler (#815)
