+++ packages/server/src/adapters/express.ts
@@ -1,8 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import type * as express from 'express';
-import { CreateContextFn, CreateContextFnOptions } from '../http';
-import { requestHandler } from '../http';
-import { BaseHandlerOptions } from '../internals/BaseHandlerOptions';
+import { CreateContextFnOptions, requestHandler } from '../http';
+import { HTTPHandlerOptions } from '../http/internals/HTTPHandlerOptions';
 import { AnyRouter } from '../router';
 
 export type CreateExpressContextOptions = CreateContextFnOptions<
@@ -10,16 +9,8 @@ export type CreateExpressContextOptions = CreateContextFnOptions<
   express.Response
 >;
 
-export type CreateExpressContextFn<TRouter extends AnyRouter> = CreateContextFn<
-  TRouter,
-  express.Request,
-  express.Response
->;
-
 export function createExpressMiddleware<TRouter extends AnyRouter>(
-  opts: {
-    createContext: CreateExpressContextFn<TRouter>;
-  } & BaseHandlerOptions<TRouter, express.Request>,
+  opts: HTTPHandlerOptions<TRouter, express.Request, express.Response>,
 ): express.Handler {
   return (req, res) => {
     const endpoint = req.path.substr(1);
